Use new GDS API Adapters behaviour
[Trello card](https://trello.com/c/B5z6Dq8q/)

Some old behaviour has been deprecated in [gds-api-adapters](https://github.com/alphagov/gds-api-adapters) so we are updating this app to comply with this before it becomes the default. Firstly, a `GdsApi::NotFound` error is now raised when the server returns a 404 or 410.  Secondly, requests now return a hash instead of an OpenStruct.  We have now opted in to this new behaviour. We are also now handling exceptions raised by gds-api-adapters in the `Scheme` model.